+++ HotelTamagotchi.Web/Controllers/TamagotchiController.cs
@@ -26,6 +26,7 @@ namespace HotelTamagotchi.Web.Controllers
         // GET: Tamagotchi
         public ActionResult Index()
         {
+
             return View(TamagotchiRepo.GetAll());
         }
 
@@ -47,6 +48,11 @@ namespace HotelTamagotchi.Web.Controllers
         // GET: Tamagotchi/Create
         public ActionResult Create()
         {
+            if (Session["User"] == null)
+            {
+                TempData["NotLoggedIn"] = "Please login or register to continue!";
+                return RedirectToAction("Index", "Home");
+            }
             return View();
         }
 
@@ -57,6 +63,16 @@ namespace HotelTamagotchi.Web.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "Id,HotelRoomId,Name,Age,Pennies,Level,Health,Boredom,Alive")] TamagotchiViewModel tamagotchi)
         {
+            if (Session["User"] == null)
+            {
+                TempData["NotLoggedIn"] = "Please login or register to continue!";
+                return RedirectToAction("Index", "Home");
+            }
+            if (!Session["Role"].Equals(UserRole.Customer))
+            {
+                TempData["NotAuthorized"] = "You are not authorized to perform these actions!";
+                return RedirectToAction("Index", "Home");
+            }
             if (ModelState.IsValid)
             {
                 TamagotchiRepo.Add(tamagotchi);
@@ -69,6 +85,16 @@ namespace HotelTamagotchi.Web.Controllers
         // GET: Tamagotchi/Edit/5
         public ActionResult Edit(int? id)
         {
+            if (Session["User"] == null)
+            {
+                TempData["NotLoggedIn"] = "Please login or register to continue!";
+                return RedirectToAction("Index", "Home");
+            }
+            if (!Session["Role"].Equals(UserRole.Customer))
+            {
+                TempData["NotAuthorized"] = "You are not authorized to perform these actions!";
+                return RedirectToAction("Index", "Home");
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -88,6 +114,16 @@ namespace HotelTamagotchi.Web.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "Id,HotelRoomId,Name,Age,Pennies,Level,Health,Boredom,Alive")] TamagotchiViewModel tamagotchi)
         {
+            if (Session["User"] == null)
+            {
+                TempData["NotLoggedIn"] = "Please login or register to continue!";
+                return RedirectToAction("Index", "Home");
+            }
+            if (!Session["Role"].Equals(UserRole.Customer))
+            {
+                TempData["NotAuthorized"] = "You are not authorized to perform these actions!";
+                return RedirectToAction("Index", "Home");
+            }
             if (ModelState.IsValid)
             {
                 TamagotchiRepo.SetChanged(tamagotchi);
@@ -99,6 +135,16 @@ namespace HotelTamagotchi.Web.Controllers
         // GET: Tamagotchi/Delete/5
         public ActionResult Delete(int? id)
         {
+            if (Session["User"] == null)
+            {
+                TempData["NotLoggedIn"] = "Please login or register to continue!";
+                return RedirectToAction("Index", "Home");
+            }
+            if (!Session["Role"].Equals(UserRole.Customer))
+            {
+                TempData["NotAuthorized"] = "You are not authorized to perform these actions!";
+                return RedirectToAction("Index", "Home");
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -116,6 +162,16 @@ namespace HotelTamagotchi.Web.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
+            if (Session["User"] == null)
+            {
+                TempData["NotLoggedIn"] = "Please login or register to continue!";
+                return RedirectToAction("Index", "Home");
+            }
+            if (!Session["Role"].Equals(UserRole.Customer))
+            {
+                TempData["NotAuthorized"] = "You are not authorized to perform these actions!";
+                return RedirectToAction("Index", "Home");
+            }
             TamagotchiViewModel tamagotchi = TamagotchiRepo.Find(id);
             TamagotchiRepo.Remove(tamagotchi);
             return RedirectToAction("Index");
